conversion/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
blinker==1.8.2
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.2
filelock==3.14.0
Flask==3.0.3
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib-metadata==6.8.0
intel-openmp==2021.4.0
ipykernel==6.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.0
jupyter_core==5.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
nest-asyncio==1.5.8
numpy==1.24.4
orjson==3.10.3
packaging==23.2
parso==0.8.3
pillow==10.3.0
platformdirs==4.0.0
prompt-toolkit==3.0.40
psutil==5.9.6
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
pyzmq==25.1.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tbb==2021.12.0
tornado==6.3.3
traitlets==5.13.0
typer==0.12.3
typing_extensions==4.8.0
ujson==5.9.0
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.9
websockets==12.0
Werkzeug==3.0.3
zipp==3.17.0
